+++ src/phijax/models/__init__.py
@@ -1,4 +1,6 @@
+from phijax.models.activations import Activation, ActivationName
 from phijax.models.contracts import InitializedModel, ModelFactory
+from phijax.models.initialization import InitializationName, Initializer
 from phijax.models.layers import FactorizedDense, PeriodicFeatures, RandomFourierFeatures
 from phijax.models.mlp import MLP, build_mlp
 from phijax.models.modified_mlp import ModifiedMLP, build_modified_mlp
@@ -8,8 +10,12 @@ from phijax.models.summary import tabulate_nnx_model
 
 __all__ = [
     "MLP",
+    "Activation",
+    "ActivationName",
     "FactorizedDense",
+    "InitializationName",
     "InitializedModel",
+    "Initializer",
     "ModelFactory",
     "ModifiedMLP",
     "PeriodicFeatures",